rerouted nav for reviews, added func to add rating to db

+++ BlockMe/Models/ListingUser.swift
@@ -15,8 +15,8 @@ struct ListingUser: Identifiable, Codable {
   var venmoHandle: String
   var phoneNumber: String
   var profileImageURL: String
-  var rating: Int
-  
+  var rating: Float
+
   enum CodingKeys: String, CodingKey {
     case id
     case firstName
